+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/Wand.java
@@ -35,7 +35,7 @@ import com.elmakers.mine.bukkit.utilities.borrowed.ConfigurationNode;
 
 public class Wand implements CostReducer {
 	protected final static int inventorySize = 27;
-	protected final static int inventoryOrganizeSize = 16;
+	protected final static int inventoryOrganizeSize = 20;
 	protected final static int inventoryOrganizeNewGroupSize = 8;
 	protected final static int hotbarSize = 9;
 	
@@ -1792,14 +1792,9 @@ public class Wand implements CostReducer {
 	}
 	
 	public void organizeInventory() {
-		if (activePlayer == null) return;
-		if (!isInventoryOpen()) return;
-		
 		WandOrganizer organizer = new WandOrganizer(this);
 		organizer.organize();
-		
 		openInventoryPage = 0;
-		updateInventory();
 		saveState();
 	}
 	
@@ -1817,5 +1812,9 @@ public class Wand implements CostReducer {
 	
 	public int getEffectColor() {
 		return effectColor;
+	}
+	
+	public Inventory getHotbar() {
+		return hotbar;
 	}
 }